test passed

+++ src/lib.rs
@@ -11,5 +11,5 @@ pub fn mul(a: i32, b: i32) -> i32 {
 }
 
 pub fn div(a: f32, b: f32) -> f32 {
-    (a / b) as f32
+    a / b
 }
